requirements.txt
python-telegram-bot
requests
FPDF
PyPDF2
reportlab
pillow